+++ changelog.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+1.1.0
+-----
+- Fix cursor position in Android TextWatcher
+- Added custom `.toString()` to all classes
+
 1.0.0
 -----
 - Library working